Added New AI
Added in UI that allows the player when he grabs the letter in the first and second level. He is able to read where the next level will be set.

Then also implemented some new code that allows me to be able to make the letters trigger when the player goes thourgh each level.